+++ docs/user/release_notes.rst
@@ -10,6 +10,7 @@ Dependency changes
 
 * Pungi 4.1.20 or higher is now required.
 * ``six`` is now a required dependency.
+* Skopeo is now a required dependency for Bodhi installations that compose containers.
 
 
 Bugs